# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "Smm Api Documentation",
        "contact": {
            "email": "takahiroasiro@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login/local": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Local login",
                "description": "Локальный логин: login/email + password",
                "operationId": "localLogin",
                "parameters": [
                    {
                        "name": "login",
                        "in": "query",
                        "description": "The login parameter in query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email parameter in query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password parameter in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
...
        "/api/update_password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Update password",
                "description": "Обновить пароль",
                "operationId": "updatePassword",
                "parameters": [
                    {
                        "name": "old_password",
                        "in": "query",
                        "description": "The old_password parameter array in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password parameter array in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirm",
                        "in": "query",
                        "description": "The password_confirm parameter array in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password changed"
                    }
                }
            }
        },
        "/api/update_avatar": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Update avatar",
                "description": "Обновить аватар",
                "operationId": "updateAvatar",
                "parameters": [
                    {
                        "name": "avatar",
                        "in": "query",
                        "description": "The avatar parameter mimes: jpeg, jpg, png",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "avatar changed"
                    }
                }
            }
        },
...
        "/api/confirm/{confirmationCode}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify registration",
                "description": "Подтверждает регистрацию нового пользователя В случае успеха устанавливает пользователю группу ROLE_VERIFIED",
                "operationId": "confirm",
                "parameters": [
                    {
                        "name": "confirmation_code",
                        "in": "path",
                        "description": "The confirmation_code parameter array in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
...
        "/api/set_password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Set password",
                "description": "Устанавливает новый пароль по коду из письма",
                "operationId": "setPassword",
                "parameters": [
                    {
                        "name": "reset_code",
                        "in": "query",
                        "description": "The reset_code parameter array in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password parameter array in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirm",
                        "in": "query",
                        "description": "The password_confirm parameter array in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}